refactor: fixing text color bug when changing theme ✒️ Previously, in the auth pages, when the user changed the color theme to light, text colors went black making it hard to see. Now its fixed!
